refactor: add entry_points package
