+++ feeluown/app.py
@@ -1,11 +1,12 @@
 import asyncio
 import logging
+import sys
 from functools import partial
 from contextlib import contextmanager
 
 from fuocore import LiveLyric, Library
 from fuocore.dispatch import Signal
-from fuocore.pubsub import run as run_pubsub
+from fuocore.pubsub import run as run_pubsub, create as create_pubsub
 
 from .consts import APP_ICON
 from .player import Player
@@ -14,7 +15,6 @@ from .server import FuoServer
 from .publishers import LiveLyricPublisher
 from .request import Request
 from .version import VersionManager
-from .fuoexec import fuoexec_after_app_attrs_attached
 
 logger = logging.getLogger(__name__)
 
@@ -26,23 +26,18 @@ class App:
     GuiMode = 0x0010     # 显示 GUI
     CliMode = 0x0100     # 命令行模式
 
-    instance = None
-
-    def exec_(self, code):
-        """执行 Python 代码"""
-        obj = compile(code, '<string>', 'single')
-        self._g.update({
-            'app': self,
-            'player': self.player
-        })
-        exec(obj, self._g, self._g)
+    def __init__(self, config):
+        self.mode = config.MODE  # DEPRECATED: use app.config.MODE instead
+        self.config = config
+        self.initialized = Signal()
 
     def show_msg(self, msg, *args, **kwargs):
         """在程序中显示消息，一般是用来显示程序当前状态"""
+        # pylint: disable=no-self-use, unused-argument
         logger.info(msg)
 
     @contextmanager
-    def create_action(self, s):
+    def create_action(self, s):  # pylint: disable=no-self-use
         """根据操作描述生成 Action (alpha)
 
         设计缘由：用户需要知道目前程序正在进行什么操作，进度怎么样，
@@ -72,12 +67,6 @@ class App:
         else:
             show_msg(s + '...done')  # done
 
-    def shutdown(self):
-        if self.mode & App.DaemonMode:
-            self.pubsub_server.close()
-        self.player.stop()
-        self.player.shutdown()
-
 
 def attach_attrs(app):
     """初始化 app 属性"""
@@ -90,13 +79,16 @@ def attach_attrs(app):
     app.playlist = app.player.playlist
     app.plugin_mgr = PluginsManager(app)
     app.request = Request()
-    app._g = {}
 
     if app.mode & (app.DaemonMode | app.GuiMode):
         app.version_mgr = VersionManager(app)
 
+    if app.mode & app.DaemonMode:
+        app.server = FuoServer(app)
+        app.pubsub_gateway, app.pubsub_server = create_pubsub()
+        app._ll_publisher = LiveLyricPublisher(app.pubsub_gateway)
+
     if app.mode & app.GuiMode:
-        from feeluown.widgets.collections import CollectionsModel
         from feeluown.uimodels.provider import ProviderUiManager
         from feeluown.uimodels.playlist import PlaylistUiManager
         from feeluown.uimodels.my_music import MyMusicUiManager
@@ -128,36 +120,22 @@ def attach_attrs(app):
         app.show_msg = app.ui.magicbox.show_msg
 
 
-def initialize(app):
-    loop = asyncio.get_event_loop()
-    app.player.position_changed.connect(app.live_lyric.on_position_changed)
-    app.playlist.song_changed.connect(app.live_lyric.on_song_changed)
-    app.plugin_mgr.scan()
-    if app.mode & app.DaemonMode:
-        app.server = FuoServer(app, loop=loop)
-        loop.create_task(app.server.run())
-        app.pubsub_gateway, app.pubsub_server = run_pubsub()
-        app._ll_publisher = LiveLyricPublisher(app.pubsub_gateway)
-        app.live_lyric.sentence_changed.connect(app._ll_publisher.publish)
-
-    if app.mode & App.GuiMode:
-        app.theme_mgr.autoload()
-        app.tips_mgr.show_random_tip()
-        app.coll_uimgr.initialize()
-
-    if app.mode & (App.DaemonMode | App.GuiMode):
-        loop.call_later(10, partial(loop.create_task, app.version_mgr.check_release()))
-    app.initialized.emit(app)
-
-
 def create_app(config):
-    bases = [App]
-
     mode = config.MODE
+
     if mode & App.GuiMode:
+
+        from quamash import QEventLoop
         from PyQt5.QtCore import QSize
         from PyQt5.QtGui import QIcon, QPixmap
         from PyQt5.QtWidgets import QApplication, QWidget
+
+        q_app = QApplication(sys.argv)
+        q_app.setQuitOnLastWindowClosed(True)
+        q_app.setApplicationName('FeelUOwn')
+
+        app_event_loop = QEventLoop(q_app)
+        asyncio.set_event_loop(app_event_loop)
 
         class GuiApp(QWidget):
             mode = App.GuiMode
@@ -168,39 +146,87 @@ def create_app(config):
                 QApplication.setWindowIcon(QIcon(QPixmap(APP_ICON)))
 
             def closeEvent(self, e):
-                app.ui.mpv_widget.close()
+                self.ui.mpv_widget.close()
                 event_loop = asyncio.get_event_loop()
                 event_loop.stop()
-                # try:
-                #     self.shutdown()
-                # finally:
-                #     QApplication.quit()
 
-            def sizeHint(self):
+            def sizeHint(self):  # pylint: disable=no-self-use
                 return QSize(1000, 618)
 
+        class FApp(App, GuiApp):
+            def __init__(self, config):
+                App.__init__(self, config)
+                GuiApp.__init__(self)
 
-        bases.append(GuiApp)
+    else:
+        FApp = App
 
-    if mode & App.CliMode:
-
-        class CliApp:
-            pass
-
-        bases.append(CliApp)
-
-    class FApp(*bases):
-        def __init__(self, mode):
-            for base in bases:
-                base.__init__(self)
-            self.mode = mode
-            self.initialized = Signal()
-    app = FApp(mode)
-    App.instance = app
-    app.config = config
+    Signal.setup_aio_support()
+    app = FApp(config)
     attach_attrs(app)
-    fuoexec_after_app_attrs_attached(app)
-    if app.mode & App.GuiMode:
-        app.show()
-    initialize(app)
     return app
+
+
+def init_app(app):
+    app.player.position_changed.connect(app.live_lyric.on_position_changed)
+    app.playlist.song_changed.connect(app.live_lyric.on_song_changed)
+    if app.mode & app.DaemonMode:
+        app.live_lyric.sentence_changed.connect(app._ll_publisher.publish)
+
+    app.plugin_mgr.scan()
+    if app.mode & App.GuiMode:
+        app.theme_mgr.autoload()
+        app.tips_mgr.show_random_tip()
+        app.coll_uimgr.initialize()
+        app.show()
+
+
+def run_app(app):
+    loop = asyncio.get_event_loop()
+
+    if app.mode & (App.DaemonMode | App.GuiMode):
+        loop.call_later(10, partial(loop.create_task, app.version_mgr.check_release()))
+
+    if app.mode & App.DaemonMode:
+        if sys.platform.lower() == 'darwin':
+            try:
+                from .global_hotkey_mac import MacGlobalHotkeyManager
+            except ImportError as e:
+                logger.warning("Can't start mac hotkey listener: %s", str(e))
+            else:
+                mac_global_hotkey_mgr = MacGlobalHotkeyManager()
+                mac_global_hotkey_mgr.start()
+        loop.create_task(app.server.run())
+        run_pubsub(app.pubsub_gateway, app.pubsub_server)
+
+    try:
+        if not (app.config.MODE & (App.GuiMode | App.DaemonMode)):
+            logger.warning('Fuo running with no daemon and no window')
+        loop.run_forever()
+    except KeyboardInterrupt:
+        # NOTE: gracefully shutdown?
+        pass
+    finally:
+        loop.stop()
+        _shutdown_app(app)
+        loop.close()
+
+
+def run_app_once(app, future):
+    loop = asyncio.get_event_loop()
+
+    try:
+        loop.run_until_complete(future)
+    except KeyboardInterrupt:
+        pass
+    finally:
+        loop.stop()
+        _shutdown_app(app)
+        loop.close()
+
+
+def _shutdown_app(app):
+    if app.mode & App.DaemonMode:
+        app.pubsub_server.close()
+    app.player.stop()
+    app.player.shutdown()
