+++ feeluown/entry_points/run_cli.py
@@ -1,0 +1,5 @@
+from feeluown.cli import climain
+
+
+def run_cli(args):
+    climain(args)
