+++ feeluown/server.py
@@ -8,12 +8,13 @@ logger = logging.getLogger(__name__)
 
 
 class FuoServer:
-    def __init__(self, app, loop):
+    def __init__(self, app):
         self._app = app
-        self._loop = loop
+        self._loop = None
 
     async def run(self, host='0.0.0.0', port=23333):
         loop = asyncio.get_event_loop()
+        self._loop = loop
         await loop.create_server(self.protocol_factory, host, port)
         logger.info('Fuo daemon run at {}:{}'.format(host, port))
